+++ src/source/geojson_source.test.ts
@@ -530,7 +530,7 @@ describe('GeoJSONSource#getBounds', () => {
                             ]
                         }
                     }]
-                })
+                });
             }
         }), undefined);
         source.map = mapStub;